FIX: use signed loop variables for OpenMP
git-svn-id: https://www.nitrc.org/svn/cmtk/trunk@3096 42a5c34f-2066-0410-bec5-ba365beb4995

+++ core/libs/Registration/cmtkTypedArraySimilarity.cxx
@@ -242,7 +242,7 @@ TypedArraySimilarity::GetCrossCorrelation
 
   unsigned int numberOfPixels = array0->GetDataSize();
 #pragma omp parallel for reduction(+:sumOfProducts,sumOfSquares0,sumOfSquares1)
-  for ( unsigned int idx = 0; idx < numberOfPixels; ++idx ) 
+  for ( int idx = 0; idx < static_cast<int>( numberOfPixels ); ++idx ) 
     {
     Types::DataItem pixel0, pixel1;
     if ( array0->Get( pixel0, idx ) && array1->Get( pixel1, idx ) )
